Working on (not done):
The builder needs to access information about the model and needs to return a CompositionBuilder<MusicView>.

The MusicReader (the controller)'s parseFile takes a readable (which the text file should be) and a CompositionBuilder<MusicView> which provides a way to build the composition, e.g. set the tempo and add a bunch of notes (since MusicReader's don't know about models). With that information we can create all the views.

The main method runs the program with environment variables.

Once all of this ^^^ is actually implemented we should be able to print song files to console from main